backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3

# AI and NLP
langchain==0.0.350

# Browser Automation
playwright==1.40.0
selenium==4.15.2

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Communication
twilio==8.10.0
elevenlabs==0.2.26

# Media Processing
ffmpeg-python==0.2.0
moviepy==1.0.3

# HTTP and Async
httpx==0.25.2
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1